Started working on optimization
Reworking the way all the hero modifications are reset when switching hero. Main goal is to reduce the number of events that are set to check conditions for all players regardless of their hero. There are also some events that are supposed to be checked for all players, but are sometimes executing unnecessarily. Hopefully can reduce the number of crashes that seem to have started occurring.