Symmetra portal changes and Mercy change
Improved the collision checking on the symmetra portal and made it so that everyone who used the portals will be teleported back in bounds after the ult is over. Also made Mercy give an extra life to her whole team at the beginning of her ultimate because the healing detection doesn't work the same way anymore, I actually already made this change before, but somehow it wasn't saved.