Started working on making all the ults work in FFA
Might be a fair amount of work tbh.

So far I've made sure that Ana, D.Va, Doomfist, Moira and Junkrat work in FFA. Baptiste and Brigitte were disabled altogether, because Baptiste's ult creates an enemy shield in FFA and with Brigitte it seems like when respawning in FFA the player's inputs will be blocked for a longer duration which kinda messed with her whole shield mechanic and I didn't have the energy to fix it.