McCree and Mei changes
McCree should work in ffa now. Mei already worked pretty well for FFA though I'm fixing some general issues with her. Put the freezing effect on a loop so that it will be set as soon as the player can be frozen. Also gonna fix some stuff with multiple Meis in the match, still WIP.