Fixed some resets applying unintentionally when switching character, FFA stuff
Characters like Tracer were resetting when switching character and making the other character slow, because the disable vanilla ultimate rule sets B to the player hero which could then cause the ultimate to be reset as well. I set some conditions in the resets to check if the characters hadn't already stopped using their ultimates.

Started working on FFA symmetra, but got bored and left it for later.

Set up Widow, Zarya and Wrecking Ball for FFA.