New Bastion ultimate WIP, Widowmaker changes
New Bastion Ultimate that fires a bunch of homing missiles in turret form.

Changed Widowmaker to utilize projectiles for better headshot detection. Added a backup projectile in case the raycasts miss, but the projectile might still hit. The hit detection seems to work best if you lead the shot a little.